sched: Ensure cpu_power periodic update
With a lot of small task, the softirq sched is nearly never called
when no_hz is enable. Te load_balance is mainly called with
the newly_idle mode which doesn't update the cpu_power.
Add a next_update field which ensure a maximum update period when
there is short activity

Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>

commit fd46ac7f3809555918399d1b49aae05260b61fd3 1 parent ed93ca452c
Vincent Guittot authored a year ago