AndreiLux: exynos-mem: Add fimc1 to mmap whitelist check
The assumption that only devices which use CONFIG_CMA_DMA use
exynos-mem is wrong. Remove that check.

Newer camera libraries as on Note 2 4.1.2 variants also map fimc directly,
but they are not reserved through kernel CMA allocation in the device file
but through Samsung's S5P framework.

There, the block is called "fimc1", so add that to the whitelist check.

commit 81c95f6046880be48ef377ebae4e42c791f0813e 1 parent a23e6f62a0
AndreiLux authored a day ago