AndreiLux: cma: Create a tool to create checks against cma block memory spaces
Introduce a frameork to collect the device's CMA memory blocks, and
functions able to check a given memory space against the blocks to see
if the accessed space given is a valid CMA memory request.

Use this to refuse access to unauthorized addresses given to the mmap
funtions in the mem-exynos and s5p-smem char devices.

commit b9c4dd16a2eae1ee8fb39f00b84531b61904191e 1 parent 468be45260
AndreiLux authored 3 days ago

Thx again!